Refactor method-invocation expression parsing
